Make integer settings use IntegerListPreference
Avoids unnecessary type casting and duplication in resource files

+++ app/src/main/java/emu/skyline/SkylineApplication.kt
@@ -16,6 +16,6 @@ class SkylineApplication : Application() {
     override fun onCreate() {
         super.onCreate()
         System.loadLibrary("skyline")
-        initializeLog(applicationContext.filesDir.canonicalPath + "/", getSettings().logLevel.toInt())
+        initializeLog(applicationContext.filesDir.canonicalPath + "/", getSettings().logLevel)
     }
 }
